requirements.txt
fastapi
uvicorn[standard]
breeze-connect
python-dotenv
pydantic
loguru


